+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Types of changes:
 - `<SectionHelper/>` - add box-sizing to not break styles [#2202](https://github.com/wix/wix-style-react/pull/2202)
 - `<Tooltip/>` - fix tooltip jumping [#2225](https://github.com/wix/wix-style-react/pull/2225)
 - `<ButtonLayout/>` - remove border from button focus [#2234](https://github.com/wix/wix-style-react/pull/2234)
+- `<InputWithOptions/>` - fix inability to re-select the same option after typing something 
 
 ### Changed
 - `<ButtonLayout/>` - update typography [#2198](https://github.com/wix/wix-style-react/pull/2198)
@@ -32,6 +33,7 @@ Types of changes:
 
 ## Removed
 - Remove dead code `src/Backoffice/ButtonLayout` (was never in use) [#2231](https://github.com/wix/wix-style-react/pull/2231)
+- `<InputWithOptions/>` - remove `selectedId` prop (it's an internal state)
 
 ### Docs
 - `<Badge/>` - migrate story to autodocs [#2221](https://github.com/wix/wix-style-react/pull/2221)